fix(ui): remove activity feed cache split dependencies
