+++ openmetadata-ui/src/main/resources/ui/src/components/ActivityFeed/ActivityFeedPanel/TaskPanelHeader.tsx
@@ -19,7 +19,7 @@ import { useTranslation } from 'react-i18next';
 import { TASK_ENTITY_TYPES } from '../../../constants/Task.constant';
 import { Task, TaskEntityStatus } from '../../../rest/tasksAPI';
 import { getEntityName } from '../../../utils/EntityNameUtils';
-import { getTaskDisplayId } from '../../../utils/TaskNavigationUtils';
+import { getTaskDisplayId } from '../../../utils/TasksUtils';
 import './feed-panel-header.less';
 
 interface TaskPanelHeaderProps {
