+++ openmetadata-ui/src/main/resources/ui/src/components/ActivityFeed/ActivityFeedProvider/ActivityFeedProvider.test.tsx
@@ -32,9 +32,7 @@ import {
   removeActivityReaction,
 } from '../../../rest/feedsAPI';
 import { listMyVisibleTasks, listTasks } from '../../../rest/tasksAPI';
-import ActivityFeedProvider, {
-  clearActivityFeedCache,
-} from './ActivityFeedProvider';
+import ActivityFeedProvider from './ActivityFeedProvider';
 import {
   DummyActivityCommentComponent,
   DummyActivityFeedComponent,
@@ -156,7 +154,6 @@ jest.mock('../../../utils/FeedUtilsPure', () => ({
 describe('ActivityFeedProvider', () => {
   beforeEach(() => {
     jest.clearAllMocks();
-    clearActivityFeedCache();
     mockUseDomainStore.mockImplementation((selector) =>
       selector({ activeDomain: 'All Domains' })
     );
