+++ openmetadata-ui/src/main/resources/ui/src/components/NotificationBox/NotificationFeedCard.component.tsx
@@ -25,7 +25,7 @@ import { entityDisplayName, prepareFeedLink } from '../../utils/FeedUtilsPure';
 import {
   getTaskDetailPathFromTask,
   getTaskDisplayId,
-} from '../../utils/TaskNavigationUtils';
+} from '../../utils/TasksUtils';
 import { ActivityFeedTabs } from '../ActivityFeed/ActivityFeedTab/ActivityFeedTab.interface';
 import ProfilePicture from '../common/ProfilePicture/ProfilePicture';
 import { SourceType } from '../SearchedData/SearchedData.interface';
