+++ openmetadata-ui/src/main/resources/ui/src/components/common/RichTextEditor/TaskDescriptionPreviewer.test.tsx
@@ -10,7 +10,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-import { act, render, screen } from '@testing-library/react';
+import { act, render, screen, waitFor } from '@testing-library/react';
 import { PreviewerProp } from './RichTextEditor.interface';
 import TaskDescriptionPreviewer from './TaskDescriptionPreviewer';
 
@@ -56,12 +56,12 @@ describe('TaskDescriptionPreviewer', () => {
     }));
   });
 
-  it('should render the component with markdown content', () => {
+  it('should render the component with markdown content', async () => {
     render(<TaskDescriptionPreviewer {...mockProp} />);
 
     expect(screen.getByTestId('viewer-container')).toBeInTheDocument();
     expect(screen.getByTestId('markdown-parser')).toBeInTheDocument();
-    expect(screen.getByTestId('block-editor')).toBeInTheDocument();
+    expect(await screen.findByTestId('block-editor')).toBeInTheDocument();
   });
 
   it('should render no-description placeholder when markdown is empty', () => {
@@ -185,12 +185,12 @@ describe('TaskDescriptionPreviewer', () => {
     });
   });
 
-  it('should update content when markdown prop changes', () => {
+  it('should update content when markdown prop changes', async () => {
     const { rerender } = render(
       <TaskDescriptionPreviewer {...mockProp} markdown="Initial content" />
     );
 
-    expect(screen.getByTestId('block-editor')).toHaveTextContent(
+    expect(await screen.findByTestId('block-editor')).toHaveTextContent(
       'Initial content'
     );
 
@@ -198,7 +198,7 @@ describe('TaskDescriptionPreviewer', () => {
       <TaskDescriptionPreviewer {...mockProp} markdown="Updated content" />
     );
 
-    expect(screen.getByTestId('block-editor')).toHaveTextContent(
+    expect(await screen.findByTestId('block-editor')).toHaveTextContent(
       'Updated content'
     );
   });
@@ -239,17 +239,19 @@ describe('TaskDescriptionPreviewer', () => {
     });
   });
 
-  it('should set BlockEditor to non-editable mode', () => {
+  it('should set BlockEditor to non-editable mode', async () => {
     const BlockEditor = require('../../BlockEditor/BlockEditor');
     render(<TaskDescriptionPreviewer {...mockProp} />);
 
-    expect(BlockEditor).toHaveBeenCalledWith(
-      expect.objectContaining({
-        editable: false,
-        autoFocus: false,
-      }),
-      {}
-    );
+    await waitFor(() => {
+      expect(BlockEditor).toHaveBeenCalledWith(
+        expect.objectContaining({
+          editable: false,
+          autoFocus: false,
+        }),
+        {}
+      );
+    });
   });
 
   it('should handle empty content after formatting', async () => {
